// services/libs/integrations/src/integrations/git/types.ts
export enum GitActivityType {
  AUTHORED_COMMIT = 'authored-commit',
  REVIEWED_COMMIT = 'reviewed-commit',
  TESTED_COMMIT = 'tested-commit',
  CO_AUTHORED_COMMIT = 'co-authored-commit',
  INFORMED_COMMIT = 'informed-commit',
  INFLUENCED_COMMIT = 'influenced-commit',
  APPROVED_COMMIT = 'approved-commit',
  COMMITTED_COMMIT = 'committed-commit',
  REPORTED_COMMIT = 'reported-commit',
  RESOLVED_COMMIT = 'resolved-commit',
  SIGNED_OFF_COMMIT = 'signed-off-commit',
}